+++ Predicate.h
@@ -10,7 +10,7 @@ public:
     Predicate(std::string h, vector<Parameter*> p){head = h; params = p;}
     void toString(){
         cout << head << "(";
-        for (int i=0; i < params.size(); i++){
+        for (size_t i=0; i < params.size(); i++){
             if(i != 0)
                 cout << ",";
             params[i]->toString();
